requirements.txt
mido
python-rtmidi
unidecode
asyncio
math
scipy
apscheduler
winrt